tapnode: hoist node-side interfaces out of tapgarden
ChainBridge, WalletAnchor, KeyRing, and GroupFetcher describe tapd's
view of its node-side dependencies (chain backend, wallet, key ring,
group lookups), but they lived in tapgarden/interface.go. They were
consumed by tapfreighter, tapchannel, tapconfig, lndservices,
universe/supplycommit, universe/supplyverifier, and backup -- which
made the tapgarden home a documented accident of authorship rather
than a property of what the package is.

Introduce a tapnode package that houses the four interfaces alone.
Concrete implementations remain in lndservices; the rest of the tree
imports tapnode for the interface only. tapgarden retains the
minting-specific symbols (MintingStore, MintSupplyCommitter, Planter,
BatchState, FundBatchResp, etc.). No behavior change.
